KA test: compare cell centers at ulps of the data type, not of Float64
cmp_float took eps(Float64(one(eltype(h)))), so Float32 centers were held to
4 ulp of Float64 (8.9e-16); CUDA's FMA contraction differs from the host by
one Float32 ulp (1.17e-7) and failed on every CUDA run since. Metal passed
only because both arms there are bit-identical.

+++ test/metal_env/ka_tree_vs_cpu_correctness.jl
@@ -136,7 +136,7 @@ function cmp_float(label, h, d, name; ulps=4)
     scale = max.(abs.(Float64.(h)), 1e-30)
     rel = dif ./ scale
     worst = isempty(rel) ? 0.0 : maximum(rel)
-    tol = ulps * eps(Float64(one(eltype(h))))
+    tol = ulps * eps(one(eltype(h)))   # ulps of the DATA type, not of Float64
     ok = worst <= tol
     ok || (nfail[] += 1;
         @printf("    FAIL %-14s max rel %.3e > %.3e (%d ulp)\n", name, worst, tol, ulps))
